requirements.txt
openai
python-dotenv
rich
PyPDF2
python-docx
